refactor: split storage record and product model roots
Break session-product DDL, runtime storage records, and archive product entities into narrower families while preserving the public import surfaces.
