+++ polylogue/archive_product_entities.py
@@ -2,342 +2,23 @@
 
 from __future__ import annotations
 
-from typing import Any
-
-from polylogue.archive_product_base import (
-    ARCHIVE_PRODUCT_CONTRACT_VERSION,
-    ArchiveEnrichmentProvenance,
-    ArchiveInferenceProvenance,
-    ArchiveProductModel,
-    ArchiveProductProvenance,
+from polylogue.archive_product_governance_entities import (
+    ArchiveDebtProduct,
+    MaintenanceRunProduct,
 )
-from polylogue.archive_product_payloads import (
-    ArchiveDebtTargetLineage,
-    SessionEnrichmentPayload,
-    SessionEvidencePayload,
-    SessionInferencePayload,
-    SessionPhaseEvidencePayload,
-    SessionPhaseInferencePayload,
-    WorkEventEvidencePayload,
-    WorkEventInferencePayload,
+from polylogue.archive_product_session_entities import (
+    SessionEnrichmentProduct,
+    SessionPhaseProduct,
+    SessionProfileProduct,
+    SessionWorkEventProduct,
+    WorkThreadProduct,
 )
-from polylogue.maintenance_models import ArchiveDebtStatus
-from polylogue.storage.store import (
-    MaintenanceRunRecord,
-    SessionPhaseRecord,
-    SessionProfileRecord,
-    SessionWorkEventRecord,
-    WorkThreadRecord,
+from polylogue.archive_product_summary_entities import (
+    DaySessionSummaryProduct,
+    ProviderAnalyticsProduct,
+    SessionTagRollupProduct,
+    WeekSessionSummaryProduct,
 )
-
-
-class SessionProfileProduct(ArchiveProductModel):
-    contract_version: int = ARCHIVE_PRODUCT_CONTRACT_VERSION
-    product_kind: str = "session_profile"
-    semantic_tier: str = "merged"
-    conversation_id: str
-    provider_name: str
-    title: str | None = None
-    provenance: ArchiveProductProvenance
-    evidence: SessionEvidencePayload | None = None
-    inference_provenance: ArchiveInferenceProvenance | None = None
-    inference: SessionInferencePayload | None = None
-
-    @classmethod
-    def from_record(
-        cls,
-        record: SessionProfileRecord,
-        *,
-        tier: str = "merged",
-    ) -> SessionProfileProduct:
-        include_evidence = tier in {"merged", "evidence"}
-        include_inference = tier in {"merged", "inference"}
-        return cls(
-            semantic_tier=tier,
-            conversation_id=record.conversation_id,
-            provider_name=record.provider_name,
-            title=record.title,
-            provenance=ArchiveProductProvenance(
-                materializer_version=record.materializer_version,
-                materialized_at=record.materialized_at,
-                source_updated_at=record.source_updated_at,
-                source_sort_key=record.source_sort_key,
-            ),
-            evidence=(
-                SessionEvidencePayload.model_validate(record.evidence_payload)
-                if include_evidence
-                else None
-            ),
-            inference_provenance=(
-                ArchiveInferenceProvenance(
-                    materializer_version=record.materializer_version,
-                    materialized_at=record.materialized_at,
-                    source_updated_at=record.source_updated_at,
-                    source_sort_key=record.source_sort_key,
-                    inference_version=record.inference_version,
-                    inference_family=record.inference_family,
-                )
-                if include_inference
-                else None
-            ),
-            inference=(
-                SessionInferencePayload.model_validate(record.inference_payload)
-                if include_inference
-                else None
-            ),
-        )
-
-
-class SessionEnrichmentProduct(ArchiveProductModel):
-    contract_version: int = ARCHIVE_PRODUCT_CONTRACT_VERSION
-    product_kind: str = "session_enrichment"
-    semantic_tier: str = "enrichment"
-    conversation_id: str
-    provider_name: str
-    title: str | None = None
-    provenance: ArchiveProductProvenance
-    enrichment_provenance: ArchiveEnrichmentProvenance
-    enrichment: SessionEnrichmentPayload
-
-    @classmethod
-    def from_record(cls, record: SessionProfileRecord) -> SessionEnrichmentProduct:
-        return cls(
-            conversation_id=record.conversation_id,
-            provider_name=record.provider_name,
-            title=record.title,
-            provenance=ArchiveProductProvenance(
-                materializer_version=record.materializer_version,
-                materialized_at=record.materialized_at,
-                source_updated_at=record.source_updated_at,
-                source_sort_key=record.source_sort_key,
-            ),
-            enrichment_provenance=ArchiveEnrichmentProvenance(
-                materializer_version=record.materializer_version,
-                materialized_at=record.materialized_at,
-                source_updated_at=record.source_updated_at,
-                source_sort_key=record.source_sort_key,
-                enrichment_version=record.enrichment_version,
-                enrichment_family=record.enrichment_family,
-            ),
-            enrichment=SessionEnrichmentPayload.model_validate(record.enrichment_payload),
-        )
-
-
-class SessionWorkEventProduct(ArchiveProductModel):
-    contract_version: int = ARCHIVE_PRODUCT_CONTRACT_VERSION
-    product_kind: str = "session_work_event"
-    semantic_tier: str = "inference"
-    event_id: str
-    conversation_id: str
-    provider_name: str
-    event_index: int
-    provenance: ArchiveProductProvenance
-    inference_provenance: ArchiveInferenceProvenance
-    evidence: WorkEventEvidencePayload
-    inference: WorkEventInferencePayload
-
-    @classmethod
-    def from_record(cls, record: SessionWorkEventRecord) -> SessionWorkEventProduct:
-        return cls(
-            event_id=record.event_id,
-            conversation_id=record.conversation_id,
-            provider_name=record.provider_name,
-            event_index=record.event_index,
-            provenance=ArchiveProductProvenance(
-                materializer_version=record.materializer_version,
-                materialized_at=record.materialized_at,
-                source_updated_at=record.source_updated_at,
-                source_sort_key=record.source_sort_key,
-            ),
-            inference_provenance=ArchiveInferenceProvenance(
-                materializer_version=record.materializer_version,
-                materialized_at=record.materialized_at,
-                source_updated_at=record.source_updated_at,
-                source_sort_key=record.source_sort_key,
-                inference_version=record.inference_version,
-                inference_family=record.inference_family,
-            ),
-            evidence=WorkEventEvidencePayload.model_validate(record.evidence_payload),
-            inference=WorkEventInferencePayload.model_validate(record.inference_payload),
-        )
-
-
-class SessionPhaseProduct(ArchiveProductModel):
-    contract_version: int = ARCHIVE_PRODUCT_CONTRACT_VERSION
-    product_kind: str = "session_phase"
-    semantic_tier: str = "inference"
-    phase_id: str
-    conversation_id: str
-    provider_name: str
-    phase_index: int
-    provenance: ArchiveProductProvenance
-    inference_provenance: ArchiveInferenceProvenance
-    evidence: SessionPhaseEvidencePayload
-    inference: SessionPhaseInferencePayload
-
-    @classmethod
-    def from_record(cls, record: SessionPhaseRecord) -> SessionPhaseProduct:
-        return cls(
-            phase_id=record.phase_id,
-            conversation_id=record.conversation_id,
-            provider_name=record.provider_name,
-            phase_index=record.phase_index,
-            provenance=ArchiveProductProvenance(
-                materializer_version=record.materializer_version,
-                materialized_at=record.materialized_at,
-                source_updated_at=record.source_updated_at,
-                source_sort_key=record.source_sort_key,
-            ),
-            inference_provenance=ArchiveInferenceProvenance(
-                materializer_version=record.materializer_version,
-                materialized_at=record.materialized_at,
-                source_updated_at=record.source_updated_at,
-                source_sort_key=record.source_sort_key,
-                inference_version=record.inference_version,
-                inference_family=record.inference_family,
-            ),
-            evidence=SessionPhaseEvidencePayload.model_validate(record.evidence_payload),
-            inference=SessionPhaseInferencePayload.model_validate(record.inference_payload),
-        )
-
-
-class WorkThreadProduct(ArchiveProductModel):
-    contract_version: int = ARCHIVE_PRODUCT_CONTRACT_VERSION
-    product_kind: str = "work_thread"
-    thread_id: str
-    root_id: str
-    dominant_project: str | None = None
-    provenance: ArchiveProductProvenance
-    thread: dict[str, Any]
-
-    @classmethod
-    def from_record(cls, record: WorkThreadRecord) -> WorkThreadProduct:
-        return cls(
-            thread_id=record.thread_id,
-            root_id=record.root_id,
-            dominant_project=record.dominant_project,
-            provenance=ArchiveProductProvenance(
-                materializer_version=record.materializer_version,
-                materialized_at=record.materialized_at,
-                source_updated_at=record.end_time or record.start_time,
-                source_sort_key=None,
-            ),
-            thread=dict(record.payload),
-        )
-
-
-class SessionTagRollupProduct(ArchiveProductModel):
-    contract_version: int = ARCHIVE_PRODUCT_CONTRACT_VERSION
-    product_kind: str = "session_tag_rollup"
-    tag: str
-    conversation_count: int
-    explicit_count: int
-    auto_count: int
-    provider_breakdown: dict[str, int]
-    project_breakdown: dict[str, int]
-    provenance: ArchiveProductProvenance
-
-
-class DaySessionSummaryProduct(ArchiveProductModel):
-    contract_version: int = ARCHIVE_PRODUCT_CONTRACT_VERSION
-    product_kind: str = "day_session_summary"
-    date: str
-    provenance: ArchiveProductProvenance
-    summary: dict[str, Any]
-
-
-class WeekSessionSummaryProduct(ArchiveProductModel):
-    contract_version: int = ARCHIVE_PRODUCT_CONTRACT_VERSION
-    product_kind: str = "week_session_summary"
-    iso_week: str
-    provenance: ArchiveProductProvenance
-    summary: dict[str, Any]
-
-
-class MaintenanceRunProduct(ArchiveProductModel):
-    contract_version: int = ARCHIVE_PRODUCT_CONTRACT_VERSION
-    product_kind: str = "maintenance_run"
-    maintenance_run_id: str
-    executed_at: str
-    mode: str
-    preview: bool
-    repair_selected: bool
-    cleanup_selected: bool
-    vacuum_requested: bool
-    target_names: tuple[str, ...] = ()
-    success: bool
-    schema_version: int
-    manifest: dict[str, Any]
-
-    @classmethod
-    def from_record(cls, record: MaintenanceRunRecord) -> MaintenanceRunProduct:
-        return cls(
-            maintenance_run_id=record.maintenance_run_id,
-            executed_at=record.executed_at,
-            mode=record.mode,
-            preview=record.preview,
-            repair_selected=record.repair_selected,
-            cleanup_selected=record.cleanup_selected,
-            vacuum_requested=record.vacuum_requested,
-            target_names=record.target_names,
-            success=record.success,
-            schema_version=record.schema_version,
-            manifest=dict(record.manifest),
-        )
-
-
-class ProviderAnalyticsProduct(ArchiveProductModel):
-    contract_version: int = ARCHIVE_PRODUCT_CONTRACT_VERSION
-    product_kind: str = "provider_analytics"
-    provider_name: str
-    conversation_count: int
-    message_count: int
-    user_message_count: int
-    assistant_message_count: int
-    avg_messages_per_conversation: float
-    avg_user_words: float
-    avg_assistant_words: float
-    tool_use_count: int
-    thinking_count: int
-    total_conversations_with_tools: int
-    total_conversations_with_thinking: int
-    tool_use_percentage: float
-    thinking_percentage: float
-
-
-class ArchiveDebtProduct(ArchiveProductModel):
-    contract_version: int = ARCHIVE_PRODUCT_CONTRACT_VERSION
-    product_kind: str = "archive_debt"
-    debt_name: str
-    category: str
-    maintenance_target: str
-    destructive: bool
-    issue_count: int
-    healthy: bool
-    detail: str
-    governance_stage: str
-    lineage: ArchiveDebtTargetLineage | None = None
-
-    @classmethod
-    def from_status(
-        cls,
-        status: ArchiveDebtStatus,
-        *,
-        governance_stage: str,
-        lineage: ArchiveDebtTargetLineage | None = None,
-    ) -> ArchiveDebtProduct:
-        return cls(
-            debt_name=status.name,
-            category=status.category.value,
-            maintenance_target=status.maintenance_target,
-            destructive=status.destructive,
-            issue_count=status.issue_count,
-            healthy=status.healthy,
-            detail=status.detail,
-            governance_stage=governance_stage,
-            lineage=lineage,
-        )
-
 
 __all__ = [
     "ArchiveDebtProduct",
