+++ polylogue/archive_product_summary_entities.py
@@ -1,0 +1,58 @@
+"""Aggregate archive product entities."""
+
+from __future__ import annotations
+
+from typing import Any
+
+from polylogue.archive_product_base import (
+    ARCHIVE_PRODUCT_CONTRACT_VERSION,
+    ArchiveProductModel,
+    ArchiveProductProvenance,
+)
+
+
+class SessionTagRollupProduct(ArchiveProductModel):
+    contract_version: int = ARCHIVE_PRODUCT_CONTRACT_VERSION
+    product_kind: str = "session_tag_rollup"
+    tag: str
+    conversation_count: int
+    explicit_count: int
+    auto_count: int
+    provider_breakdown: dict[str, int]
+    project_breakdown: dict[str, int]
+    provenance: ArchiveProductProvenance
+
+
+class DaySessionSummaryProduct(ArchiveProductModel):
+    contract_version: int = ARCHIVE_PRODUCT_CONTRACT_VERSION
+    product_kind: str = "day_session_summary"
+    date: str
+    provenance: ArchiveProductProvenance
+    summary: dict[str, Any]
+
+
+class WeekSessionSummaryProduct(ArchiveProductModel):
+    contract_version: int = ARCHIVE_PRODUCT_CONTRACT_VERSION
+    product_kind: str = "week_session_summary"
+    iso_week: str
+    provenance: ArchiveProductProvenance
+    summary: dict[str, Any]
+
+
+class ProviderAnalyticsProduct(ArchiveProductModel):
+    contract_version: int = ARCHIVE_PRODUCT_CONTRACT_VERSION
+    product_kind: str = "provider_analytics"
+    provider_name: str
+    conversation_count: int
+    message_count: int
+    user_message_count: int
+    assistant_message_count: int
+    avg_messages_per_conversation: float
+    avg_user_words: float
+    avg_assistant_words: float
+    tool_use_count: int
+    thinking_count: int
+    total_conversations_with_tools: int
+    total_conversations_with_thinking: int
+    tool_use_percentage: float
+    thinking_percentage: float
